+++ packages/core/crates/core-network/src/messaging.rs
@@ -19,6 +19,7 @@ impl ControlMessage {
     pub fn generate_ping_request() -> Self {
         let mut ping = PingMessage::default();
         ping.nonce.copy_from_slice(&derive_ping_pong(None));
+
         Self::Ping(ping)
     }