+++ packages/core/crates/core-types/src/acknowledgement.rs
@@ -10,7 +10,7 @@ use utils_types::primitives::Address;
 use utils_types::traits::BinarySerializable;
 
 /// Represents packet acknowledgement
-#[derive(Clone, Debug, PartialEq)]
+#[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
 #[cfg_attr(feature = "wasm", wasm_bindgen::prelude::wasm_bindgen(getter_with_clone))]
 pub struct Acknowledgement {
     ack_signature: OffchainSignature,
@@ -275,7 +275,7 @@ pub mod test {
     use crate::acknowledgement::{AcknowledgedTicket, Acknowledgement, PendingAcknowledgement, UnacknowledgedTicket};
     use crate::channels::Ticket;
     use core_crypto::keypairs::{ChainKeypair, Keypair, OffchainKeypair};
-    use core_crypto::types::{Challenge, CurvePoint, HalfKey, Hash, OffchainPublicKey, Response};
+    use core_crypto::types::{Challenge, CurvePoint, HalfKey, OffchainPublicKey, Response};
     use ethnum::u256;
     use hex_literal::hex;
     use utils_types::primitives::{Address, Balance, BalanceType, U256};