+++ packages/core/src/channel-strategy.ts
@@ -15,12 +15,8 @@ import {
   StrategyTickResult,
   Balance,
   BalanceType,
-  utils_misc_initialize_crate
-} from '../lib/core_strategy.js'
-
-utils_misc_initialize_crate()
-
-export { StrategyTickResult } from '../lib/core_strategy.js'
+  PeerQuality
+} from '../lib/core_hopr.js'
 
 import { ChannelStatus, AcknowledgedTicket, ChannelEntry } from '@hoprnet/hopr-utils'
 import HoprCoreEthereum from '@hoprnet/hopr-core-ethereum'
@@ -51,12 +47,7 @@ export interface ChannelStrategyInterface {
 
   configure(settings: any): void
 
-  tick(
-    balance: BN,
-    network_addresses: Iterator<string>,
-    outgoing_channel: OutgoingChannelStatus[],
-    peer_quality: (string: string) => number
-  ): StrategyTickResult
+  tick(balance: BN, network_addresses: PeerQuality, outgoing_channel: OutgoingChannelStatus[]): StrategyTickResult
 
   onChannelWillClose(channel: ChannelEntry): Promise<void> // Before a channel closes
   onAckedTicket(t: AcknowledgedTicket): Promise<void>
@@ -116,9 +107,8 @@ export abstract class SaneDefaults {
 interface RustStrategyInterface {
   tick: (
     balance: Balance,
-    network_addresses: Iterator<string>,
-    outgoing_channels: OutgoingChannelStatus[],
-    peer_quality: (string: string) => number
+    network_addresses: PeerQuality,
+    outgoing_channels: OutgoingChannelStatus[]
   ) => StrategyTickResult
   name: string
 }
@@ -140,18 +130,8 @@ class RustStrategyWrapper<T extends RustStrategyInterface> extends SaneDefaults 
     this.autoRedeemTickets = settings.auto_redeem_tickets ?? false
   }
 
-  tick(
-    balance: BN,
-    network_addresses: Iterator<string>,
-    outgoing_channels: OutgoingChannelStatus[],
-    peer_quality: (string: string) => number
-  ): StrategyTickResult {
-    return this.strategy.tick(
-      new Balance(balance.toString(), BalanceType.HOPR),
-      network_addresses,
-      outgoing_channels,
-      peer_quality
-    )
+  tick(balance: BN, network_addresses: PeerQuality, outgoing_channels: OutgoingChannelStatus[]): StrategyTickResult {
+    return this.strategy.tick(new Balance(balance.toString(), BalanceType.HOPR), network_addresses, outgoing_channels)
   }
 
   get name() {