+++ packages/core/src/network.ts
@@ -1,22 +1,21 @@
 import {
   resolve_network,
-  core_misc_initialize_crate,
+  core_hopr_initialize_crate,
   supported_networks,
   type ChainOptions,
   type ResolvedNetwork,
   type Network
-} from '../lib/core_misc.js'
-core_misc_initialize_crate()
+} from '../lib/core_hopr.js'
+core_hopr_initialize_crate()
 
 import type { DeploymentExtract } from '@hoprnet/hopr-core-ethereum'
-
 export {
   resolve_network,
   supported_networks,
   type ChainOptions,
   type ResolvedNetwork,
   type Network
-} from '../lib/core_misc.js'
+} from '../lib/core_hopr.js'
 
 export type EnvironmentType = 'production' | 'staging' | 'development' | 'local'