+++ packages/core/src/types.ts
@@ -1,1 +1,0 @@
-export { Acknowledgement } from '../../core/lib/core_types.js'